Authentication and Registration Controllers
Moved register from the user controller to a registration controller. Moved login and logout from the user controller to a authentication controller.
