+++ VotingApplication/Controllers/UserController.cs
@@ -90,174 +90,8 @@ namespace VotingApplication.Controllers
             return View("ForgotPassword/CheckEmail");
         }
 
-        #region Registration
-        [HttpGet]
-        [AllowAnonymous]
-        public IActionResult Registration(string returnUrl = null)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-
-            return View("Registration/Index"); //Index view
-        }
-
-        [HttpPost]
-        [AllowAnonymous]
-        public async Task<IActionResult> Registration(RegistrationViewModel model, string returnUrl = null)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-
-            if (ModelState.IsValid)
-            {
-                var user = new ApplicationUser
-                {
-                    UserName = model.Username,
-                    Email = model.Email
-                };
-
-                var result = await _UserManager.CreateAsync(user, model.Password);
-
-                if (result.Succeeded)
-                {
-                    var fromAddress = new MailAddress("votingnotification6@gmail.com", "Voting System");
-                    var toAddress = new MailAddress(model.Email, model.Username);
-                    const string fromPassword = "Team6Admin";
-                    const string subject = "Welcome Confirmation Email";
-                    string body = "Hello " + model.Username;
-
-                    var smtp = new SmtpClient
-                    {
-                        Host = "smtp.gmail.com",
-                        Port = 587,
-                        EnableSsl = true,
-                        DeliveryMethod = SmtpDeliveryMethod.Network,
-                        UseDefaultCredentials = false,
-                        Credentials = new NetworkCredential(fromAddress.Address, fromPassword)
-                    };
-                    using (var message = new MailMessage(fromAddress, toAddress)
-                    {
-                        Subject = subject,
-                        Body = body
-                    })
-                    {
-                        smtp.Send(message);
-                    }
-
-                    await _SignInManager.SignInAsync(user, isPersistent: false);
-
-                    return RedirectToAction(nameof(Dashboard));
-                }
-            }
-            
-            return View("Registration/Index", model);
-        }
-
-        [AcceptVerbs("Get", "Post")]
-        public async Task<IActionResult> VerifyPassword(string password)
-        {
-            if (_UserManager != null)
-            {
-                string errors = "";
-
-                IdentityResult result;
-
-                foreach (var validator in _UserManager.PasswordValidators)
-                {
-                    result = await validator.ValidateAsync(_UserManager, null, password);
-                    foreach (var error in result.Errors)
-                    {
-                        errors += error.Description + "\n";
-                    }
-                }
-                if (errors != string.Empty)
-                {
-                    return Json($"{errors}");
-                }
-            }
-
-            return Json(true);
-        }
-
-        [AcceptVerbs("Get", "Post")]
-        public async Task<IActionResult> VerifyUser(string username, string email)
-        {
-            if (_UserManager != null)
-            {
-                string errorTarget = username == null ? "Email" : "UserName";
-
-                var user = new ApplicationUser
-                {
-                    UserName = username,
-                    Email = email
-                };
-
-                string errors = "";
-
-                IdentityResult result;
-
-                foreach (var validator in _UserManager.UserValidators)
-                {
-                    result = await validator.ValidateAsync(_UserManager, user);
-                    
-                    foreach(var error in result.Errors)
-                    {
-                        if(error.Code.Contains(errorTarget))
-                            errors += error.Description + "\n";
-                    }
-                }
-                if (errors != string.Empty)
-                {
-                    return Json($"{errors}");
-                }
-            }
-
-            return Json(true);
-        }
-        #endregion
-
-        #region Login
-        [HttpGet]
-        [AllowAnonymous]
-        public IActionResult Login(string returnUrl = null)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-
-            return View("Login/Index"); //Index view
-        }
-
-        [HttpPost]
-        [AllowAnonymous]
-        public async Task<IActionResult> Login(LoginViewModel model, string returnUrl = null)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-
-            await HttpContext.SignOutAsync(IdentityConstants.ApplicationScheme);
-
-            if (ModelState.IsValid)
-            {
-                var result = await _SignInManager.PasswordSignInAsync(model.Username, model.Password, isPersistent: true, lockoutOnFailure: false);
-
-                if (result.Succeeded)
-                {
-                    if (string.IsNullOrEmpty(returnUrl))
-                        return RedirectToAction(nameof(Dashboard));
-
-                    return Redirect(returnUrl);
-                }
-                else
-                {
-                    ModelState.AddModelError(string.Empty, "Invalid username or password.");
-                }
-            }
-
-            return View("Login/Index", model);
-        }
-        #endregion
-
-        public async Task<IActionResult> LogoutAsync()
-        {
-            await HttpContext.SignOutAsync(IdentityConstants.ApplicationScheme);
-
-            return RedirectToAction(nameof(Login));
-        }
+        
+        
+        
     }
 }
