+++ VotingApplication/Views/Registration/Register.cshtml
@@ -1,9 +1,4 @@
 ﻿@model RegistrationViewModel
-@{
-    // move this code into the C# class so it is easier to edit
-    ViewData["Purpose"] = "Create Account";
-    ViewData["Submit"] = "Create";
-}
 
 <script>
     $.validator.setDefaults({
@@ -11,7 +6,7 @@
     })
 </script>
 
-<form asp-route-returnUrl="@ViewData["ReturnUrl"]" method="post" autocomplete="off">
+<form asp-action="RegisterAsync" asp-route-returnUrl="@ViewData["ReturnUrl"]" method="post" autocomplete="off">
     <h2 class="form-signin-heading">@ViewData["Purpose"]</h2>
     <div class="form-group">
         <label asp-for="Username"></label>
@@ -38,4 +33,4 @@
     </div>
     <button type="submit" class="btn btn-primary">@ViewData["Submit"]</button>
 </form>
-<a asp-area="" asp-controller="User" asp-action="Login" class="nav-link">Already have an account?</a>
+<a asp-area="" asp-controller="Authentication" asp-action="Login" class="nav-link">Already have an account?</a>
