+++ VotingApplication/Views/Shared/_Layout.cshtml
@@ -30,16 +30,16 @@
                         <a asp-area="" asp-controller="User" asp-action="Dashboard" class="nav-link">Dashboard</a>
                     </li>
                     <li class="nav-item ">
-                        <a asp-area="" asp-controller="User" asp-action="Login" class="nav-link">Login</a>
+                        <a asp-area="" asp-controller="Authentication" asp-action="Login" class="nav-link">Login</a>
                     </li>
                     <li class="nav-item ">
-                        <a asp-area="" asp-controller="User" asp-action="Registration" class="nav-link">Create Account</a>
+                        <a asp-area="" asp-controller="Registration" asp-action="Register" class="nav-link">Create Account</a>
                     </li>
                     <li class="nav-item ">
                         <a asp-area="" asp-controller="User" asp-action="Profile" class="nav-link">Profile</a>
                     </li>
                     <li class="nav-item ">
-                        <a asp-area="" asp-controller="User" asp-action="LogoutAsync" class="nav-link">Logout</a>
+                        <a asp-area="" asp-controller="Authentication" asp-action="LogoutAsync" class="nav-link">Logout</a>
                     </li>
                 </ul>
             </div>
